pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roadsense-ai"
version = "0.1.0"
description = "Agentic RAG platform for roadside assistance and insurance claim triage."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "httpx>=0.27",
  "structlog>=24.1",
  "redis>=5.0",
  "langgraph>=0.2",
  "llama-index-core>=0.10",
  "pydantic-ai>=0.0.13",
  "azure-functions>=1.19",
  "azure-functions-durable>=1.2",
  "presidio-analyzer>=2.2",
  "presidio-anonymizer>=2.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
  "mypy>=1.10",
]
eval = [
  "ragas>=0.1.14",
  "datasets>=2.20",
]

[tool.setuptools]
packages = ["api", "api.routes", "agents", "agents.tools", "ingestion", "ingestion.activities", "observability", "security"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
